feat: moviegen, reset and started work on text
